fix: address PR review comments for validation and broadening
- profiles.py: use dtype=float in zeros_like to prevent integer grid failures
- emissivity.py: validate scalar sigma_nm is finite and positive before broadening
- test_physics_invariants.py: fix inconsistent .get() default in ionization test

+++ cflibs/radiation/profiles.py
@@ -127,7 +127,7 @@ def apply_gaussian_broadening_per_line(
     if len(sigmas) > 0 and np.any(np.asarray(sigmas) <= 0):
         raise ValueError("All per-line sigma values must be > 0")
 
-    spectrum = np.zeros_like(wavelength_grid)
+    spectrum = np.zeros_like(wavelength_grid, dtype=float)
 
     for wl, intensity, sig in zip(line_wavelengths, line_intensities, sigmas):
         profile = gaussian_profile(wavelength_grid, wl, sig, intensity)
